Rewriting the test-mock.
Created Regatta class in common. This holds the ID, Name and UTcOffset for a regatta.
Created classes to generate their respective class (Reggatta, Race, Course).
These classes implement the Generator Interface which currently has a single method called generate. This has been implemented only in RegattaGenerator so far.

Rewriting so that the test-mock works in the following way:
1. Create necessary classes (Regatta, Course, Race) where attributes are randomised by using the appropriate Generator class.
2. Pass these classes to another class (for example, something like: AC35XMLGenerator). This will then split the data up into the appropriate XML files for the AC35 protocol.