[Just re-comitting as I forgot a # in the message previously]
Finish splitting the reading of XML files and generating race, course classes from them out of the TestMock class.

Created enum TestXMLFiles which holds file paths to the regatta, race, boat xml files that get sent to clients and used for the simulation.
Fixed all of the tests so that they work with the split.

#story[1062]